pii-engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
presidio-analyzer>=2.2.355
presidio-anonymizer>=2.2.355
spacy>=3.8.0
pdfplumber>=0.11.0
python-docx>=1.1.0
reportlab>=4.2.0
python-multipart>=0.0.9
